# Optimize data structure usage a bit