--- a/Fluid.Tests/CaseStatementTests.cs
+++ b/Fluid.Tests/CaseStatementTests.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using Fluid.Ast;
@@ -14,9 +15,9 @@ namespace Fluid.Tests
         private LiteralExpression C = new LiteralExpression(new StringValue("c"));
         private LiteralExpression D = new LiteralExpression(new StringValue("d"));
 
-        private Statement[] TEXT(string text)
+        private List<Statement> TEXT(string text)
         {
-            return new Statement[] { new TextStatement(text) };
+            return new List<Statement> { new TextStatement(text) };
         }
 
         [Fact]
@@ -26,7 +27,7 @@ namespace Fluid.Tests
                 A,
                 null,
                 new[] {
-                    new WhenStatement(new [] { A }, TEXT("x"))
+                    new WhenStatement(new List<Expression> { A }, TEXT("x"))
                 }
             );
 
@@ -43,7 +44,7 @@ namespace Fluid.Tests
                 A,
                 null,
                 new[] {
-                    new WhenStatement(new [] { A, B, C }, TEXT("x"))
+                    new WhenStatement(new List<Expression> { A, B, C }, TEXT("x"))
                 }
             );
 
@@ -60,7 +61,7 @@ namespace Fluid.Tests
                 A,
                 null,
                 new[] {
-                    new WhenStatement(new [] { B, C, D }, TEXT("x"))
+                    new WhenStatement(new List<Expression> { B, C, D }, TEXT("x"))
                 }
             );
 
@@ -75,9 +76,9 @@ namespace Fluid.Tests
         {
             var e = new CaseStatement(
                 A,
-                new ElseStatement(new[] { new TextStatement("y")}),
+                new ElseStatement(new List<Statement> { new TextStatement("y")}),
                 new[] {
-                    new WhenStatement(new [] { A }, TEXT("x"))
+                    new WhenStatement(new List<Expression> { A }, TEXT("x"))
                 }
             );
 
@@ -92,9 +93,9 @@ namespace Fluid.Tests
         {
             var e = new CaseStatement(
                 A,
-                new ElseStatement(new[] { new TextStatement("y") }),
+                new ElseStatement(new List<Statement> { new TextStatement("y") }),
                 new[] {
-                    new WhenStatement(new [] { B, C }, TEXT("x"))
+                    new WhenStatement(new List<Expression> { B, C }, TEXT("x"))
                 }
             );
 
@@ -109,11 +110,11 @@ namespace Fluid.Tests
         {
             var e = new CaseStatement(
                 B,
-                new ElseStatement(new[] { new TextStatement("y") }),
+                new ElseStatement(new List<Statement> { new TextStatement("y") }),
                 new[] {
-                    new WhenStatement(new [] { A, C }, TEXT("1")),
-                    new WhenStatement(new [] { B, C }, TEXT("2")),
-                    new WhenStatement(new [] { C }, TEXT("3"))
+                    new WhenStatement(new List<Expression> { A, C }, TEXT("1")),
+                    new WhenStatement(new List<Expression> { B, C }, TEXT("2")),
+                    new WhenStatement(new List<Expression> { C }, TEXT("3"))
                 }
             );
 
@@ -130,8 +131,8 @@ namespace Fluid.Tests
                 A,
                 null,
                 new[] {
-                    new WhenStatement(new [] { B }, TEXT("2")),
-                    new WhenStatement(new [] { C }, TEXT("3"))
+                    new WhenStatement(new List<Expression> { B }, TEXT("2")),
+                    new WhenStatement(new List<Expression> { C }, TEXT("3"))
                 }
             );