--- a/Fluid/BlockContext.cs
+++ b/Fluid/BlockContext.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 using Fluid.Ast;
 using Irony.Parsing;
 
@@ -10,10 +8,10 @@ namespace Fluid
     {
         // Some types of sub-block can be repeated (when, case)
         // This value is initialized the first time a sub-block is found
-        private Dictionary<string, IList<Statement>> _blocks;
+        private Dictionary<string, List<Statement>> _blocks;
 
         // Statements that are added while parsing a sub-block (else, elsif, when)
-        private IList<Statement> _transientStatements;
+        private List<Statement> _transientStatements;
 
         public BlockContext(ParseTreeNode tag)
         {
@@ -24,13 +22,13 @@ namespace Fluid
 
         public ParseTreeNode Tag { get; }
 
-        public IList<Statement> Statements { get; private set; }
+        public List<Statement> Statements { get; }
 
         public void EnterBlock(string name, TagStatement statement)
         {
             if (_blocks == null)
             {
-                _blocks = new Dictionary<string, IList<Statement>>();
+                _blocks = new Dictionary<string, List<Statement>>();
             }
 
             if (!_blocks.TryGetValue(name, out var blockStatements))
@@ -47,14 +45,24 @@ namespace Fluid
             _transientStatements.Add(statement);
         }
 
-        public IList<TStatement> GetBlockStatements<TStatement>(string name)
+        public List<TStatement> GetBlockStatements<TStatement>(string name) where TStatement : Statement
         {
             if (_blocks != null && _blocks.TryGetValue(name, out var statements))
             {
-                return statements.Cast<TStatement>().ToList();
+                var result = new List<TStatement>(statements.Count);
+                for (int i = 0; i < statements.Count; ++i)
+                {
+                    result.Add((TStatement) statements[i]);
+                }
+                return result;
             }
 
-            return Array.Empty<TStatement>();
+            return StatementList<TStatement>.Empty;
+        }
+
+        private static class StatementList<T>
+        {
+            internal static readonly List<T> Empty = new List<T>();
         }
     }
 }