--- a/Fluid/Filters/NumberFilters.cs
+++ b/Fluid/Filters/NumberFilters.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Text;
 using Fluid.Values;
 
 namespace Fluid.Filters