* prefer concrete types
  * IList already allows mutation so List doesn't change things much, mostly showed array to IList boxing
* for is faster for list
* pool string builders

## Fluid.Benchmarks.FluidBenchmarks

| **Diff**|Method|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------:|-------:|-------:|-------:|-------|
| Old |ParseSample|40.604 us|6.4697|-|-|27152 B|
| **New** |	| **34.445 us (-15%)** | **6.4697 (0%)** | **-** | **-** | **27152 B (0%)** |
| Old |ParseAndRenderSample|52.659 us|8.4229|-|-|35515 B|
| **New** |	| **44.011 us (-16%)** | **8.1787 (-3%)** | **-** | **-** | **34483 B (-3%)** |
| Old |RenderSample|7.171 us|1.9913|-|-|8361 B|
| **New** |	| **6.685 us (-7%)** | **1.7471 (-12%)** | **-** | **-** | **7329 B (-12%)** |
| Old |ParseLoremIpsum|2.104 us|0.0839|-|-|360 B|
| **New** |	| **2.107 us (0%)** | **0.0839 (0%)** | **-** | **-** | **360 B (0%)** |
| Old |RenderSimpleOuput|6.534 us|1.6327|-|-|6864 B|
| **New** |	| **5.611 us (-14%)** | **1.5945 (-2%)** | **-** | **-** | **6720 B (-2%)** |
| Old |RenderLoremSimpleOuput|7.529 us|2.9984|-|-|12584 B|
| **New** |	| **6.489 us (-14%)** | **2.4948 (-17%)** | **-** | **-** | **10488 B (-17%)** |